requirements.txt
dotenv
asyncio
pinecone-client
pydantic
chromadb
langfuse
python-multipart
llama-index>=0.12.0
llama-index-core>=0.12.0
llama-index-tools-google>=0.3.0
llama-index-readers-file>=0.4.0
llama-index-embeddings-huggingface>=0.5.0
llama-index-llms-huggingface>=0.5.0
llama-index-llms-huggingface-api>=0.4.0
llama-index-vector-stores-chroma>=0.4.0
llama-index-vector-stores-pinecone>=0.1.0
llama-index-callbacks-arize-phoenix>=0.1.0
llama-index-embeddings-huggingface-api>=0.3.0